Support custom naming schemes for ZooKeeper groups
Added support for custom naming schemes in ZooKeeper groups that
implement the Group.NodeNameScheme interface that are passed in through
the constructor.

Added backwards compatibility in the form of Group.DefaultNamingScheme
so that existing usage does not break.

Documented changes in constructor API, the new interface, and the new
default implementation.

Added test GroupTest.testJoinCustomNamingScheme to test that the nodes
are named according to the new naming scheme and a listener is updated
regarding their presence.