Artifact support for jar_dependencies.
Proper support for ivy artifacts via jar_dependency#with_artifact.
Moves the logic to generate an artifact if theres an ext/url (and
now type or classifier) into the jar_dependency and out of the
mako template. The ext/url members of jar_dependency are kept around
for backwards compatibility both of __init__ and publishing jars.