Fix bug in analysis cache parser.
When there is a placeholder scala_library target with no sources,
the analysis cache file isn't generated by the cache split operation.
The cache file reader doesn't examine the targets at all, so it doesn't
know that there's a good reason why the cache file wasn't generated.

This change skips a cache with a warning when the cache file isn't found.
The only cases I've seen of this are empty placeholder targets, but
it's better to generate a warning rather than skipping silently.