[PATCH 3661/7508] Add support for instance Arg fields
The meat of this change is in the addition of Args.from which permits specifying
objects to use arg fields from.  This is the basic step needed to allow apps to transition
to using non-global Args which is important for testability and "re-entrant" mains.

This change also converts the 1st customer of instance args - a main that will be run in
a nailgun and thus requires re-entrant arg parsing.

In the course of implementing this change I cleaned up warnings, docs and code in
general in touched files as well as eliminating some dead code.

RB_ID=261047

(sapling split of 4ed4137ba6b9aa24a73b3aec2536817d3df6b2d5)