[PATCH 05/27] Restrict chunking to internal targets.
Senthil and I believe this was introduced in https://reviewboard.twitter.biz/r/256435 . Recently during some internal changes we switched from chunking internal targets to concrete targets. Typically this works fine, however, there's a case where generating IDEA project files with the --idea-python option does the following:

(a) The Idea task generation task constructor registers interest in python products because the user explicitly asked for that.

(b) Thrift code generation looks to see that people explicitly asked for python thrift codegen and does that. It generates code and creates synthetic PythonLibrary targets to own the sources.

(c) This is where the issue is. Due to the recent group engine change we try to iterate over all the concrete targets, which includes PythonLibrary instances. However, PythonLibrary instances are not InternalTarget subclasses and this leads to the stack trace we see below.

The proposed fix is changing this behavior back to what it was before the refactor, which is to restrict to internal targets when making chunks.

RB_ID=273399

(sapling split of 87fc9fdfbaf15b793b5a96ff75d32bc3d93f6383)