pants PythonTargets should walk provided PythonBinary artifacts
The change to .resolve for PythonLibrary to extend to its binaries in the
case it specified a PythonArtifact caused a regression with pants py.
Instead we should just augment _walk to walk those targets so that they're
properly reflected as members in setup_py.

Reviewed at https://rbcommons.com/s/twitter/r/57/