Fixup pants dev mode
This is a crutch fix only - use the full requirement set from the
pants target to bootstrap with.  This works for bootstrapping a
pants.pex since its additive and the pex creation should be hermetic
anyhow and it gives full functionality when running under PANTS_DEV=1
against the live sources.

Also align 3rdparty/python:setuptools version with common/python's default.